django/requirements/base.txt
Django==1.5.1
MySQL-python==1.2.3
Pillow==2.4.0
South==0.7.6
URLObject==2.0.2
Unipath==1.0
Werkzeug==0.8.3
amqplib==1.0.2
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.17
celery==3.0.11
django-appconf==0.5
django-audiofield==0.5.2
django-celery==3.0.11
django-celery-with-redis==3.0
django-extensions==0.9
django-imagekit==2.0.2
django-polymorphic==0.5.3
django-redis-sessions==0.3.1
django-reusableapps==0.1.1
django-tastypie==0.9.11
django-uuidfield==0.4.0
djangorestframework==0.4.0
gunicorn==0.14.6
ipdb==0.6.1
ipython==0.13.1
kombu==2.4.7
mimeparse==0.1.3
pyreadline==2.0
python-dateutil==1.5
python-gflags==2.0
python-magic==0.4.2
pytz==2012g
readline==6.2.4.1
redis==2.7.1
switch2bill-common==2.1.2
virtualenv==1.7
wsgiref==0.1.2
yolk==0.4.3
django-taggit==0.11.2
django-crispy-forms==1.4.0
django-jsonview==0.3.0
django-modeltranslation==0.7.3
Unidecode==0.04.14
django-autoslug==1.7.2
easy-thumbnails==1.5
